% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_t.R
\name{prob_t}
\alias{prob_t}
\title{Calcula a Probabilidade de t}
\usage{
prob_t(alpha, df)
}
\arguments{
\item{alpha}{Nível de significância desejado}

\item{df}{graus de liberdade da amostra}
}
\value{
Um valor entre 0 e 1.
}
\description{
Calcula a probabilidade bicaudal de z estar entre os valores críticos de um certo alpha,
considerando uma distribuição T de Student.
}
\examples{
alpha <- 0.05
prob_t(alpha, df = 10)


}
