% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_z.R
\name{prob_z}
\alias{prob_z}
\title{Calcula a Probabilidade de z}
\usage{
prob_z(alpha, mean = 0, sd = 1)
}
\arguments{
\item{alpha}{Nível de significância desejado}

\item{mean}{média da distribuição (por padrão, mean = 0)}

\item{sd}{desvio padrão da distribuição (por padrão, sd = 1)}
}
\value{
Um valor entre 0 e 1.
}
\description{
Calcula a probabilidade bicaudal de z estar entre os valores críticos de um certo alpha,
considerando uma distribuição normal.
}
\examples{
alpha <- 0.05
prob_z(alpha)


}
